F01: Add implementation of UtilisateurService and fixes ExcelController bugs
