+++ src/main/java/com/carto/sn/service/CartoImpl.java
@@ -251,49 +251,49 @@ public class CartoImpl implements ICarto{
 //		//return prof;
 //	}
 
-	@Override
-	public Utilisateur ajoutUtilisateur(Long idUtilisateur, String nom, String prenom, String login, String password, String profil) {
-		Utilisateur utilisateur = null;
-		Profil profilRole = profilRepository.findByNomProfil(profil);
-		Set<Profil> roles = Stream.of(profilRole)
-                .collect(Collectors.toCollection(HashSet::new));
-		if(idUtilisateur == null) {
-			utilisateur = utilisateurRepository.save(new Utilisateur(nom, prenom, login, passwordEncoder.encode(password), true));
-			
-		}else if(idUtilisateur!=null){
-			utilisateur = utilisateurRepository.findById(idUtilisateur).orElse(utilisateur);
-			utilisateur.setNomUtilisateur(nom);
-			utilisateur.setPrenomUtilisateur(prenom);
-			utilisateur.setLogin(login);
-			if(password.isBlank()==false)
-				utilisateur.setPassword(passwordEncoder.encode(password));
-				utilisateurRepository.save(utilisateur);
-		}
-		utilisateur.setProfil(roles);
-		
-		return utilisateur;
-	}
+//	@Override
+//	public Utilisateur ajoutUtilisateur(Long idUtilisateur, String nom, String prenom, String login, String password, String profil) {
+//		Utilisateur utilisateur = null;
+//		Profil profilRole = profilRepository.findByNomProfil(profil);
+//		Set<Profil> roles = Stream.of(profilRole)
+//                .collect(Collectors.toCollection(HashSet::new));
+//		if(idUtilisateur == null) {
+//			utilisateur = utilisateurRepository.save(new Utilisateur(nom, prenom, login, passwordEncoder.encode(password), true));
+//			
+//		}else if(idUtilisateur!=null){
+//			utilisateur = utilisateurRepository.findById(idUtilisateur).orElse(utilisateur);
+//			utilisateur.setNomUtilisateur(nom);
+//			utilisateur.setPrenomUtilisateur(prenom);
+//			utilisateur.setLogin(login);
+//			if(password.isBlank()==false)
+//				utilisateur.setPassword(passwordEncoder.encode(password));
+//				utilisateurRepository.save(utilisateur);
+//		}
+//		utilisateur.setProfil(roles);
+//		
+//		return utilisateur;
+//	}
 
 //	@Override
 //	public List <Profil> tousLesProfils() {
 //		return profilRepository.findAll();
 //	}
 
-	@Override
-	public List<Utilisateur> tousLesUtilisateurs() {
-		return utilisateurRepository.findAll();
-	}
+//	@Override
+//	public List<Utilisateur> tousLesUtilisateurs() {
+//		return utilisateurRepository.findAll();
+//	}
 
-	@Override
-	public void supprimerUtilisateur(Long idUtilisateur) {
-		 utilisateurRepository.deleteById(idUtilisateur);
-		
-	}
+//	@Override
+//	public void supprimerUtilisateur(Long idUtilisateur) {
+//		 utilisateurRepository.deleteById(idUtilisateur);
+//		
+//	}
 
-	@Override
-	public Optional<Utilisateur> findUtilisateurById(Long idUtilisateur) {
-		return utilisateurRepository.findById(idUtilisateur);
-	}
+//	@Override
+//	public Optional<Utilisateur> findUtilisateurById(Long idUtilisateur) {
+//		return utilisateurRepository.findById(idUtilisateur);
+//	}
 
 //	@Override
 //	public void supprimerRegion(Long idRegion) {
@@ -596,18 +596,18 @@ public class CartoImpl implements ICarto{
 //		return projet;
 //	}
 
-	@Override
-	public List<Utilisateur> getAllOperationsOfUser(String login) {
-		AuditReader reader = AuditReaderFactory.get(factory.createEntityManager());
-		AuditQuery queryOperationsOfUser = reader.createQuery()
-				.forRevisionsOfEntity(Utilisateur.class, true, false)
-				//.forRevisionsOfEntityWithChanges(Utilisateur.class, true)
-				.add(AuditEntity.revisionProperty("userModificateur").eq(login));
-		List <Utilisateur> listOperationsOfUser = queryOperationsOfUser.getResultList();
-		
-		return listOperationsOfUser;
-		
-	}
+//	@Override
+//	public List<Utilisateur> getAllOperationsOfUser(String login) {
+//		AuditReader reader = AuditReaderFactory.get(factory.createEntityManager());
+//		AuditQuery queryOperationsOfUser = reader.createQuery()
+//				.forRevisionsOfEntity(Utilisateur.class, true, false)
+//				//.forRevisionsOfEntityWithChanges(Utilisateur.class, true)
+//				.add(AuditEntity.revisionProperty("userModificateur").eq(login));
+//		List <Utilisateur> listOperationsOfUser = queryOperationsOfUser.getResultList();
+//		
+//		return listOperationsOfUser;
+//		
+//	}
 
 //	@Override
 //	public Profil findByIdProfil(Long idProfil) {
@@ -652,11 +652,11 @@ public class CartoImpl implements ICarto{
 //		return listPrivileges;
 //	}
 
-	@Override
-	public void enableOrDisableUserAccount(Long idUtilisateur, boolean isEnabled) {
-		Optional <Utilisateur> utilisateur = utilisateurRepository.findById(idUtilisateur);
-		utilisateur.get().setEnabled(isEnabled);
-	}
+//	@Override
+//	public void enableOrDisableUserAccount(Long idUtilisateur, boolean isEnabled) {
+//		Optional <Utilisateur> utilisateur = utilisateurRepository.findById(idUtilisateur);
+//		utilisateur.get().setEnabled(isEnabled);
+//	}
 
 //	@Override
 //	public void supprimerCategorie(Long idCategorie) {
