+++ src/main/java/com/carto/sn/service/ICarto.java
@@ -38,11 +38,11 @@ public interface ICarto {
 //	public Optional <Partenaire> findPartenaireById(Long id);
 //	public List<Projet> chercher(String motCle);
 //	public Profil ajoutProfil(String profil);
-	public Utilisateur ajoutUtilisateur(Long idUtilisateur, String nom, String prenom, String login, String password, String profil);
+//	public Utilisateur ajoutUtilisateur(Long idUtilisateur, String nom, String prenom, String login, String password, String profil);
 //	public List <Profil> tousLesProfils();
-	public List <Utilisateur> tousLesUtilisateurs();
-	public void supprimerUtilisateur(Long idUtilisateur);
-	public Optional<Utilisateur> findUtilisateurById(Long idUtilisateur);
+//	public List <Utilisateur> tousLesUtilisateurs();
+//	public void supprimerUtilisateur(Long idUtilisateur);
+//	public Optional<Utilisateur> findUtilisateurById(Long idUtilisateur);
 //	public void supprimerRegion(Long idRegion);
 //	public void supprimerProfil(Long idProfil);
 //	public void supprimerProjet(Long idProjet);
@@ -87,11 +87,11 @@ public interface ICarto {
 //	public Type findByIdType(Long idType);
 //	public Categorie findByIdCategorie(Long idCategorie);
 //	public Projet cloturerProjet(Long idProjet);
-	public List<Utilisateur> getAllOperationsOfUser(String login);
+//	public List<Utilisateur> getAllOperationsOfUser(String login);
 //	public Profil findByIdProfil(Long idProfil);
 //	public Privilege ajoutPrivilege(Long idProfil, String[] enumPrivilege);
 //	public List<Privilege> findPrivilegeByProfil(Long nomProfil);
-	public void enableOrDisableUserAccount(Long idUtilisateur, boolean isEnabled);
+//	public void enableOrDisableUserAccount(Long idUtilisateur, boolean isEnabled);
 //	public void supprimerCategorie(Long idCategorie);
 	
 }
